# docs/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meridian IT-PMO",
    "version": "5.16.0",
    "summary": "Read a governed project portfolio, and the trail of how it got that way",
    "description": "The surface other systems read. It is versioned in the path: the day the shape has to change it becomes /api/v2, so that nobody has to guess. Authentication is a key issued per connected system, carrying explicit scopes; the product never stores the key itself, only its fingerprint.\n\nWrites arrived with the first real integrator (RT365, docs/33): PUT by your own identifier on projects, milestones, register items, stage progress, work items, decisions and actions, under two write scopes, with an optional Idempotency-Key. A scope is only published once a route honours it: declaring one earlier would advertise a door that does not exist and imply it is guarded.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4173",
      "description": "A local instance"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Issued from Administration → Connected systems. `Authorization: Bearer <key>` is accepted too."
      }
    },
...
  },
  "security": [
    {
      "apiKey": []
    }
  ],
  "paths": {
    "/api/v1/actions/{externalId}": {
      "put": {
        "summary": "Create or update an action, keyed by your own identifier",
        "description": "Creates the an action when nothing carries this externalId for your integration, updates it otherwise. The identity is (integration, externalId): two connected systems may both say \"E01\" without colliding. `version` is optional — when sent it is asserted (409 if stale); when omitted your system is the master of this row and the last write wins. Optional `Idempotency-Key` header: the same key with the same body replays the recorded answer (Idempotent-Replayed: true); the same key with another body is refused (422) — one key names ONE request, not a run. `adopt: \"<Meridian id>\"` binds your externalId to a row that already exists (created on a screen, or scaffolded) instead of creating another. A field this collection does not declare is REFUSED (400) and nothing is written or audited: the properties below are the whole of what a body may carry (REQ-19). The same business rules as the screens apply. An action is raised in an OPEN meeting: send `occurrence` (an open occurrence id) or `series` (the series whose open occurrence takes it). The API never opens a meeting — a chair does.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-required-scope": "write:meetings",
        "parameters": [
          {
            "name": "externalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 200
            },
            "description": "Your own identifier for this row — stable across runs, unique within your integration"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            },
            "description": "One key names ONE request. The same key with the same body replays the recorded answer (response header Idempotent-Replayed: true); the same key with a different body is refused (422)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "adopt": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "detail": {
                    "type": "string"
                  },
                  "owner": {
                    "type": "string"
                  },
                  "project": {
                    "type": "string"
                  },
                  "dueDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "status": {
                    "type": "string"
                  },
                  "occurrence": {
                    "type": "string"
                  },
                  "series": {
                    "type": "string"
                  },
                  "version": {
                    "type": "integer"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The document described above",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contract": {
                      "type": "string"
                    },
                    "generatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "id": {
                      "type": "string"
                    },
                    "externalId": {
                      "type": "string"
                    },
                    "created": {
                      "type": "boolean"
                    },
                    "version": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Created — the row did not exist for this externalId",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contract": {
                      "type": "string"
                    },
                    "generatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "id": {
                      "type": "string"
                    },
                    "externalId": {
                      "type": "string"
                    },
                    "created": {
                      "type": "boolean"
                    },
                    "version": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The body breaks a rule the screens enforce too; the message says which"
          },
          "401": {
            "description": "No key, an unknown key, or a revoked key — the three are indistinguishable on purpose"
          },
          "403": {
            "description": "The key is live but does not carry the scope this route requires"
          },
          "409": {
            "description": "A stale version, an immutable decision, or a meeting that is not open"
          },
          "422": {
            "description": "Idempotency-Key reused with a different request"
          }
        }
      }
    },
    "/api/v1/actions": {
      "get": {
        "summary": "The actions — what a room asked of someone, and whether it is done",
        "description": "Read them before you write them: an action closed on a screen and rewritten `Open` by an unchanged re-run overwrites the minute of a meeting. `raisedInStatus` says whether the occurrence that raised it is still open. Filters: status, limit (default 200, max 500).",
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-required-scope": "read:meetings",
        "responses": {
          "200": {
            "description": "The document described above",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contract": {
                      "type": "string"
                    },
                    "generatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "actions": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No key, an unknown key, or a revoked key — the three are indistinguishable on purpose"
          },
          "403": {
            "description": "The key is live but does not carry the scope this route requires"
          }
        }
      }
    },
    "/api/v1/activities/{externalId}": {
      "put": {
        "summary": "Create or update the link to a schedule stage, and its measured progress, keyed by your own identifier",
        "description": "Creates the the link to a schedule stage, and its measured progress when nothing carries this externalId for your integration, updates it otherwise. The identity is (integration, externalId): two connected systems may both say \"E01\" without colliding. `version` is optional — when sent it is asserted (409 if stale); when omitted your system is the master of this row and the last write wins. Optional `Idempotency-Key` header: the same key with the same body replays the recorded answer (Idempotent-Replayed: true); the same key with another body is refused (422) — one key names ONE request, not a run. `adopt: \"<Meridian id>\"` binds your externalId to a row that already exists (created on a screen, or scaffolded) instead of creating another. A field this collection does not declare is REFUSED (400) and nothing is written or audited: the properties below are the whole of what a body may carry (REQ-19). The same business rules as the screens apply. Stages are not created by integrations — the plan belongs to the project. The first call binds your id to an existing stage (`activity`); every call may carry `pct`, stamped with `source` and `measuredAt` so earned value reads measured, not typed, progress (I-5).",
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-required-scope": "write:portfolio",
        "parameters": [
          {
            "name": "externalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 200
            },
            "description": "Your own identifier for this row — stable across runs, unique within your integration"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            },
            "description": "One key names ONE request. The same key with the same body replays the recorded answer (response header Idempotent-Replayed: true); the same key with a different body is refused (422)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "activity": {
                    "type": "string"
                  },
                  "pct": {
                    "type": "integer"
                  },
                  "source": {
                    "type": "string"
                  },
                  "measuredAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "name": {
                    "type": "string"
                  },
                  "version": {
                    "type": "integer"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The document described above",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contract": {
                      "type": "string"
                    },
                    "generatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "id": {
                      "type": "string"
                    },
                    "externalId": {
                      "type": "string"
                    },
                    "created": {
                      "type": "boolean"
                    },
                    "version": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Created — the row did not exist for this externalId",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contract": {
                      "type": "string"
                    },
                    "generatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "id": {
                      "type": "string"
                    },
                    "externalId": {
                      "type": "string"
                    },
                    "created": {
                      "type": "boolean"
                    },
                    "version": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The body breaks a rule the screens enforce too; the message says which"
          },
          "401": {
            "description": "No key, an unknown key, or a revoked key — the three are indistinguishable on purpose"
          },
          "403": {
            "description": "The key is live but does not carry the scope this route requires"
          },
          "409": {
            "description": "A stale version, an immutable decision, or a meeting that is not open"
          },
          "422": {
            "description": "Idempotency-Key reused with a different request"
          }
        }
      }
    },
    "/api/v1/audit": {
      "get": {
        "summary": "The audit trail — every recorded decision and change",
        "description": "Its own scope, because it is the most sensitive read in the product: a system that copies the portfolio into a warehouse has no reason to carry away who approved what. Filters: entity, entityId, action, limit, before.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-required-scope": "read:audit",
        "responses": {
          "200": {
            "description": "The document described above",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contract": {
                      "type": "string"
                    },
                    "generatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No key, an unknown key, or a revoked key — the three are indistinguishable on purpose"
          },
          "403": {
            "description": "The key is live but does not carry the scope this route requires"
          }
        }
      }
    },
...
    "/api/v1/decisions": {
      "get": {
        "summary": "The decision register — what was decided, by whom, and on what basis",
        "description": "Every decision, whether it was minuted in a meeting or taken outside one by somebody who holds the authority (I-7). Each row carries its `externalId` when an integration owns it, which is how `adopt` finds a row it wrote before it had an external identity, and its `version`, which is what an update asserts. Its own scope, mirroring `write:meetings`: a warehouse feed carries figures, not governance. Filter: limit (default 200, max 500).",
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-required-scope": "read:meetings",
        "responses": {
          "200": {
            "description": "The document described above",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contract": {
                      "type": "string"
                    },
                    "generatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "decisions": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No key, an unknown key, or a revoked key — the three are indistinguishable on purpose"
          },
          "403": {
            "description": "The key is live but does not carry the scope this route requires"
          }
        }
      }
    },
...
    "/api/v1/openapi.json": {
      "get": {
        "summary": "This contract, as served by this instance",
        "description": "The same document published at docs/openapi.v1.json, carrying the version actually running here and the address it was asked at — two things a file in a repository cannot know. Open to any live key.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-required-scope": null,
        "responses": {
          "200": {
            "description": "The document described above",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "openapi": {
                      "type": "string"
                    },
                    "info": {
                      "type": "object"
                    },
                    "servers": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "paths": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No key, an unknown key, or a revoked key — the three are indistinguishable on purpose"
          },
          "403": {
            "description": "The key is live but does not carry the scope this route requires"
          }
        }
      }
    },
    "/api/v1/portfolio": {
      "get": {
        "summary": "The whole portfolio the key is entitled to read",
        "description": "The same output the screens are drawn from — deliberately, because two projections would drift apart at the first change and an integrator would then be reading numbers nobody can see in the product. Money is in millions; dates are ISO-8601 calendar dates.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-required-scope": "read:portfolio",
        "responses": {
          "200": {
            "description": "The document described above",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contract": {
                      "type": "string"
                    },
                    "generatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "asAt": {
                      "type": "string",
                      "format": "date"
                    },
                    "counts": {
                      "type": "object"
                    },
                    "portfolio": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No key, an unknown key, or a revoked key — the three are indistinguishable on purpose"
          },
          "403": {
            "description": "The key is live but does not carry the scope this route requires"
          }
        }
      }
    },
...
  }
}